refactored all the faustPingPongDelay hierarchy of files. I made the host and index.js files generic 100%
