+++ packages/faustPingPongDelay/plugin/Gui/img/knobs/Aqua.png
