+++ packages/faustPingPongDelay/plugin/Gui/img/switches/switch_1.png
