+++ packages/faustPingPongDelay/plugin/Gui/index.js
@@ -1,9 +1,9 @@
-// This file is the factory for the GUI part..bundleRenderer.renderToStream
+// This file is the factory for the GUI part.
 // The imporant thing here is the createElement async method
 // This file must be an es module in order to be loaded with the SDK (with dynamic imports)
-import PingPongDelayFaustGui from './Gui.js';
+import Gui from './Gui.js';
 
-export { PingPongDelayFaustGui };
+export { Gui };
 
 /**
  * A mandatory method if you want a gui for your plugin
@@ -13,5 +13,5 @@ export { PingPongDelayFaustGui };
 export async function createElement(plugin, ...args) {
 	// here we return the WebComponent GUI but it could be
 // any DOM node
-	return new PingPongDelayFaustGui(plugin, ...args);
+	return new Gui(plugin, ...args);
 }
