+++ packages/faustPingPongDelay/plugin/descriptor.json
